--- src/pages/ProfilePage/components/ProfileView.tsx
@@ -1,7 +1,0 @@
-import React from "react"
-
-const ProfileView = ({ children }: React.PropsWithChildren) => {
-  return <>{children}</>
-}
-
-export default ProfileView
